feat: move iter loop logic to a separate module
